Unused Code refactored
